Fix useEffect in Actor.js
The useEffect in Actor.js was re-rendering the components in an infinite loop. This
resulted in fetchActorDetails being called infinitely. I added an empty array as an argument
to useEffect so that it renders only on initial render.
